mcp_trendscout/requirements.txt
fastmcp>=2.0.0
httpx>=0.27.0
google-adk==1.20.0
python-dotenv>=1.0.0
google-genai
asyncio
pydantic
